refactor: move all   singUp protocols to a single file
