+++ src/presentation/controllers/signup/signupProtocools.ts
@@ -1,0 +1,4 @@
+export * from "../../protocools/index";
+export * from "../../protocools/emailValidator";
+export * from "../../../domain/useCase/add-account";
+export * from "../../../domain/models/account";
