+++ src/presentation/protocools/index.ts
@@ -1,3 +1,2 @@
 export * from "./controller";
-export * from "./emailValidator";
 export * from "./http";
